改进传参，不再需要单独的模块存储
